Minor code refactoring, add sourceMaps arg to debugging
